chore: move sdkutil/coing.go
